CORE-15909: move GroupParametersLookupImpl and GroupParametersLookupInternal
