+++ components/flow/flow-service/build.gradle
@@ -15,6 +15,7 @@ dependencies {
     implementation project(":components:configuration:configuration-read-service")
     implementation project(":components:external-messaging-services")
     implementation project(":components:membership:membership-group-read")
+    implementation project(':components:membership:group-policy')
     implementation project(":components:virtual-node:virtual-node-info-read-service")
     implementation project(":components:virtual-node:cpi-info-read-service")
     implementation project(":components:virtual-node:cpk-read-service")
@@ -31,6 +32,7 @@ dependencies {
     implementation project(':libs:flows:session-manager')
     implementation project(":libs:flows:flow-utils")
     implementation project(":libs:lifecycle:lifecycle")
+    implementation project(":libs:membership:membership-common")
     implementation project(":libs:metrics")
     implementation project(":libs:messaging:messaging")
     implementation project(':libs:platform-info')
@@ -79,6 +81,7 @@ dependencies {
     integrationTestImplementation project(':testing:sandboxes')
     integrationTestImplementation project(':testing:virtual-node-info-read-service-fake')
     integrationTestImplementation project(':testing:cpi-info-read-service-fake')
+    integrationTestImplementation project(':testing:group-policy-test-common')
 
     integrationTestRuntimeOnly project(':libs:application:application-impl')
     integrationTestRuntimeOnly project(':libs:flows:session-manager-impl')
