+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/groupparameters/verifier/SignedGroupParametersVerifierImpl.kt
@@ -2,7 +2,7 @@ package net.corda.ledger.utxo.flow.impl.groupparameters.verifier
 
 import net.corda.crypto.cipher.suite.SignatureVerificationService
 import net.corda.ledger.common.data.transaction.TransactionMetadataInternal
-import net.corda.ledger.utxo.flow.impl.groupparameters.GroupParametersLookupInternal
+import net.corda.flow.application.GroupParametersLookupInternal
 import net.corda.membership.lib.SignedGroupParameters
 import net.corda.sandbox.type.SandboxConstants.CORDA_SYSTEM_SERVICE
 import net.corda.sandbox.type.UsedByFlow
