+++ libs/flows/flow-api/src/main/java/net/corda/flow/application/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.flow.application;
+
+import org.osgi.annotation.bundle.Export;
